Revert "Revert "Move shared.types to type - conflict with stdlib types under docker/py3.12""
This reverts commit 33e631ec44fa8a1877e2eb8088896fe4f6ef9a3c.
